+++ plugin-dev/Source/Sentry/Private/HAL/PlatformSentrySamplingContext.h
@@ -6,7 +6,7 @@
 
 #if PLATFORM_ANDROID
 #include "Android/AndroidSentrySamplingContext.h"
-#elif PLATFORM_APPLE
+#elif PLATFORM_APPLE && !USE_SENTRY_NATIVE
 #include "Apple/AppleSentrySamplingContext.h"
 #elif USE_SENTRY_NATIVE
 #include "GenericPlatform/GenericPlatformSentrySamplingContext.h"