+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Add out-of-process screenshot capturing on Windows ([#1325](https://github.com/getsentry/sentry-unreal/pull/1325))
+- Add native backend support for Mac ([#1333](https://github.com/getsentry/sentry-unreal/pull/1333))
 
 ### Fixes